Small fixes, testing presented bugs and now are squashed.
Makefile now has nice cc messages, flags are hidden and all formated nicely. (sucker for formality).
Database magic cheks the first 7 bytes instead of all of them. The 8th byte is not important as it was only present to be an even number of bytes.
Increased the package name and description buffer sizes.
Lined up the package list categories, spacing is fixed and looks good.